requirements.txt
flask==3.0.0
torch==2.2.1
torchvision==0.17.1
Pillow==10.0.0
requests==2.31.0
